Restructure helpers as Vue plugin
